+++ llama-starter/common/__init__.py
@@ -9,7 +9,7 @@ from common.service import RAGService
 from common.rag import SimpleRAG, load_documents, split_documents, create_rag_chain
 from common.factories import get_embeddings, get_llm, get_vectorstore
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = [
     "get_config",